Added Pulseguide
Added pulseguide to Skywatcher.

To do this, had to make some changes:
a) Implemented end close loop slew in SKywatcher.cpp to make as efficient as possible.

b) To allow this,had to restrict the possible slew rates to a maximum of siderial - prevents lots of the protection logic for e.g fast slews to be required.